unify packages rollup configuration
