+++ ui/packages/dashboard_client/rollup.config.js
@@ -1,12 +1,7 @@
-import typescript from 'rollup-plugin-typescript2'
-import commonjs from 'rollup-plugin-commonjs'
-import external from 'rollup-plugin-peer-deps-external'
-import resolve from 'rollup-plugin-node-resolve'
-
 import pkg from './package.json'
+import baseConfig from '../rollup.config.base'
 
 export default {
-  input: 'src/index.ts',
   output: [
     {
       file: pkg.main,
@@ -21,16 +16,5 @@ export default {
       sourcemap: true,
     },
   ],
-  external: ['url'],
-  plugins: [
-    external({
-      includeDependencies: true,
-    }),
-    resolve(),
-    typescript({
-      rollupCommonJSResolveHack: true,
-      clean: true,
-    }),
-    commonjs(),
-  ],
+  ...baseConfig,
 }
