+++ ui/packages/dashboard_components/package.json
@@ -36,7 +36,6 @@
     "@types/react": "^16.9.22",
     "babel-core": "^6.26.3",
     "babel-plugin-external-helpers": "^6.22.0",
-    "babel-plugin-import": "^1.13.0",
     "babel-preset-env": "^1.7.0",
     "babel-preset-react": "^6.24.1",
     "babel-preset-stage-0": "^6.24.1",
