+++ ui/packages/dashboard_components/rollup.config.js
@@ -1,23 +1,7 @@
-import typescript from 'rollup-plugin-typescript2'
-import createTransformer from 'ts-import-plugin'
-import commonjs from 'rollup-plugin-commonjs'
-import external from 'rollup-plugin-peer-deps-external'
-import postcss from 'rollup-plugin-postcss'
-import resolve from 'rollup-plugin-node-resolve'
-import url from 'rollup-plugin-url'
-import svgr from '@svgr/rollup'
-
 import pkg from './package.json'
-
-// https://github.com/Brooooooklyn/ts-import-plugin
-const transformer = createTransformer({
-  libraryDirectory: 'es',
-  libraryName: 'antd',
-  style: true,
-})
+import baseConfig from '../rollup.config.base'
 
 export default {
-  input: 'src/index.ts',
   output: [
     {
       file: pkg.main,
@@ -30,38 +14,5 @@ export default {
       sourcemap: true,
     },
   ],
-  plugins: [
-    external({
-      includeDependencies: true,
-    }),
-    postcss({
-      module: true,
-      extensions: ['.css', '.scss', '.less'],
-      use: [
-        'sass',
-        [
-          'less',
-          {
-            javascriptEnabled: true,
-            globalVars: {
-              '@padding-page': '48px', // TODO: keep same with root project
-            },
-          },
-        ],
-      ],
-    }),
-    url(),
-    svgr(),
-    resolve(),
-    typescript({
-      rollupCommonJSResolveHack: true,
-      clean: true,
-      transformers: [
-        () => ({
-          before: transformer,
-        }),
-      ],
-    }),
-    commonjs(),
-  ],
+  ...baseConfig,
 }
